Coalesce G6 survey primitives into p9-core
